Rename time_misc -> time_util
